requirements.txt
clip==0.2.0
Flask==3.1.0
flask_cors==5.0.0
matplotlib==3.10.0
numpy==2.2.1
Pillow==11.0.0
Requests==2.32.3
scikit_learn==1.6.0
torch==2.5.1
git+https://github.com/openai/CLIP.git